Create validation.js
