+++ src/components/validation.js
@@ -1,0 +1,52 @@
+const showInputError = (formElement, inputElement, errorMessage, settings) => {
+    const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+    inputElement.classList.add(settings.inputErrorClass);
+    errorElement.textContent = errorMessage;
+    errorElement.classList.add(settings.errorClass);
+  };
+  
+const hideInputError = (formElement, inputElement, settings) => {
+    const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+    inputElement.classList.remove(settings.inputErrorClass);
+    errorElement.classList.remove(settings.errorClass);
+    errorElement.textContent = '';
+};
+  
+const checkInputValidity = (formElement, inputElement, settings) => {
+    if (!inputElement.validity.valid) {
+        showInputError(formElement, inputElement, inputElement.validationMessage, settings);
+    } else {
+        hideInputError(formElement, inputElement, settings);
+    }
+};
+  
+const toggleButtonState = (inputList, buttonElement, settings) => {
+    const hasInvalidInput = inputList.some(inputElement => !inputElement.validity.valid);
+    buttonElement.disabled = hasInvalidInput;
+    buttonElement.classList.toggle(settings.inactiveButtonClass, hasInvalidInput);
+};
+  
+const setEventListeners = (formElement, settings) => {
+    const inputList = Array.from(formElement.querySelectorAll(settings.inputSelector));
+    const buttonElement = formElement.querySelector(settings.submitButtonSelector);
+
+    inputList.forEach(inputElement => {
+        inputElement.addEventListener('input', () => {
+        checkInputValidity(formElement, inputElement, settings);
+        toggleButtonState(inputList, buttonElement, settings);
+        });
+    });
+
+    toggleButtonState(inputList, buttonElement, settings);
+};
+  
+const enableValidation = (settings) => {
+    const formList = Array.from(document.querySelectorAll(settings.formSelector));
+    formList.forEach(formElement => {
+        formElement.addEventListener('submit', evt => evt.preventDefault());
+        setEventListeners(formElement, settings);
+    });
+};
+
+
+export { enableValidation };
