+++ src/scripts/index.js
@@ -1,5 +1,6 @@
 import '../pages/index.css'
 import { initialCards } from './cards.js';
+import { enableValidation } from '../components/validation.js';
 
 // Create object which contains validation classes
 const validationSettings = {
@@ -29,7 +30,6 @@ const profileDescription = document.querySelector('.profile__description');
 //Initialize profile edit popup buttons
 const profileEditButton = document.querySelector('.profile__edit-button');
 const profileEditCloseButton = document.querySelector('.popup_type_edit .popup__close');
-//const profileEditCloseButton = document.querySelector('.popup_type_edit').querySelector('.popup__close');
 
 //Initialize add card popup buttons
 const profileAddButton = document.querySelector('.profile__add-button');
@@ -47,7 +47,6 @@ profileEditButton.addEventListener('click', () => {
     profilePopup.querySelector('.popup__input_type_description').value = profileDescription.textContent;
 
     const profileEditForm = document.querySelector('.popup_type_edit .popup__form');
-    //const profileEditSubmitButton = document.querySelector('.popup_type_edit').querySelector('.popup__button');
 
     const nameInput = profileEditForm.querySelector('.popup__input_type_name');
     const jobInput = profileEditForm.querySelector('.popup__input_type_description');
@@ -103,6 +102,7 @@ cardPopup.addEventListener('submit', handleCardFormSubmit);
 profileAddButtonClose.addEventListener('click', () => {
     closeModal(currentOpenPopup);
 })
+
 
 imagePopupCloseButton.addEventListener('click', () => {
     closeModal(currentOpenPopup);
@@ -160,30 +160,31 @@ function handleOverlayClose(evt) {
  * @returns 
  */
 function createCard(card) {
-        const cardElement = cardTemplate.querySelector('.card').cloneNode(true);
-        cardElement.querySelector('.card__title').textContent = card.name;
-        cardElement.querySelector('.card__image').alt = card.name;
-        cardElement.querySelector('.card__image').src = card.link;
+  const cardElement = cardTemplate.querySelector('.card').cloneNode(true);
+  cardElement.querySelector('.card__title').textContent = card.name;
+  cardElement.querySelector('.card__image').alt = card.name;
+  cardElement.querySelector('.card__image').src = card.link;
 
-        const likeButton = cardElement.querySelector('.card__like-button');
-        likeButton.addEventListener('click', () => {
-            likeButton.classList.toggle('card__like-button_is-active');
-        })
+  const likeButton = cardElement.querySelector('.card__like-button');
+  likeButton.addEventListener('click', () => {
+      likeButton.classList.toggle('card__like-button_is-active');
+  })
 
-        const deleteButton = cardElement.querySelector('.card__delete-button');
-        deleteButton.addEventListener('click', () => {
-            cardElement.remove();
-        })
+  const deleteButton = cardElement.querySelector('.card__delete-button');
+  deleteButton.addEventListener('click', () => {
+      cardElement.remove();
+  })
 
-        const imageButton = cardElement.querySelector('.card__image');
-        imageButton.addEventListener('click', () => {
-            imagePopup.classList.add('popup_is-opened');
-            imagePopup.querySelector('.popup__image').src = card.link;
-            imagePopup.querySelector('.popup__image').alt = card.name;
-            imagePopup.querySelector('.popup__caption').textContent = card.name;
-        })
+  const imageButton = cardElement.querySelector('.card__image');
+  imageButton.addEventListener('click', () => {
+      currentOpenPopup = imagePopup;
+      imagePopup.classList.add('popup_is-opened');
+      imagePopup.querySelector('.popup__image').src = card.link;
+      imagePopup.querySelector('.popup__image').alt = card.name;
+      imagePopup.querySelector('.popup__caption').textContent = card.name;
+  })
 
-        return cardElement;
+  return cardElement;
 }
 
 
@@ -205,56 +206,4 @@ profilePopup.classList.add('popup_is-animated')
 cardPopup.classList.add('popup_is-animated')
 imagePopup.classList.add('popup_is-animated')
 
-
-// validation.js
-const showInputError = (formElement, inputElement, errorMessage, settings) => {
-    const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
-    inputElement.classList.add(settings.inputErrorClass);
-    errorElement.textContent = errorMessage;
-    errorElement.classList.add(settings.errorClass);
-  };
-  
-  const hideInputError = (formElement, inputElement, settings) => {
-    const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
-    inputElement.classList.remove(settings.inputErrorClass);
-    errorElement.classList.remove(settings.errorClass);
-    errorElement.textContent = '';
-  };
-  
-  const checkInputValidity = (formElement, inputElement, settings) => {
-    if (!inputElement.validity.valid) {
-      showInputError(formElement, inputElement, inputElement.validationMessage, settings);
-    } else {
-      hideInputError(formElement, inputElement, settings);
-    }
-  };
-  
-  const toggleButtonState = (inputList, buttonElement, settings) => {
-    const hasInvalidInput = inputList.some(inputElement => !inputElement.validity.valid);
-    buttonElement.disabled = hasInvalidInput;
-    buttonElement.classList.toggle(settings.inactiveButtonClass, hasInvalidInput);
-  };
-  
-  const setEventListeners = (formElement, settings) => {
-    const inputList = Array.from(formElement.querySelectorAll(settings.inputSelector));
-    const buttonElement = formElement.querySelector(settings.submitButtonSelector);
-  
-    inputList.forEach(inputElement => {
-      inputElement.addEventListener('input', () => {
-        checkInputValidity(formElement, inputElement, settings);
-        toggleButtonState(inputList, buttonElement, settings);
-      });
-    });
-    
-    toggleButtonState(inputList, buttonElement, settings);
-  };
-  
-  const enableValidation = (settings) => {
-    const formList = Array.from(document.querySelectorAll(settings.formSelector));
-    formList.forEach(formElement => {
-      formElement.addEventListener('submit', evt => evt.preventDefault());
-      setEventListeners(formElement, settings);
-    });
-  };
-  
 enableValidation(validationSettings);
